fix: Move new footer into layout and delete old footer
