+++ src/redesign/components/About.jsx
@@ -1,5 +1,5 @@
 import Header from "./Header";
-import Footer from "./Footer";
+import Footer from "../../layout/Footer.jsx";
 import Section from "./Section";
 import style from "../style";
 import PageHeader from "./PageHeader";
