+++ src/redesign/components/Contact.jsx
@@ -1,5 +1,5 @@
 import Header from "./Header";
-import Footer from "./Footer";
+import Footer from "../../layout/Footer";
 import Section from "./Section";
 import PageHeader from "./PageHeader";
 import style from "../style";
