+++ src/redesign/components/Testimonials.jsx
@@ -1,5 +1,5 @@
 import Header from "./Header";
-import Footer from "./Footer";
+import Footer from "../../layout/Footer";
 import Section from "./Section";
 import style from "../style";
 import { fullQuotes } from "../data/Quotes";
